[DRAFT] Cherry-pick fixes from main branch

+++ test/conformance/enqueue/urEnqueueUSMMemcpy2D.cpp
@@ -169,7 +169,7 @@ TEST_P(urEnqueueUSMMemcpy2DNegativeTest, InvalidSize) {
 TEST_P(urEnqueueUSMMemcpy2DNegativeTest, InvalidEventWaitList) {
     // enqueue something to get an event
     ur_event_handle_t event = nullptr;
-    int fill_pattern = 14;
+    uint8_t fill_pattern = 14;
     ASSERT_SUCCESS(urEnqueueUSMFill2D(queue, pDst, pitch, sizeof(fill_pattern),
                                       &fill_pattern, width, height, 0, nullptr,
                                       &event));
